Fix race condition between MergeOp and MergeabilityChecker
MergeabilityChecker is a GitReferenceUpdatedListener, which gets
called after the reference updates but before the change status
is written to the database. To reduce latency during merge the
rescan operation is sent to a background thread pool.

It is possible for the background thread to scan the database
and observe the recently merged change with status at SUBMITTED,
as the scan can run before the status is changed to MERGED.

If the MergeOp thread writes the status to the database before
Mergeable finishes the check, then Mergeable will replace the
status with SUBMITTED, making the change appear to be unsubmitted.

One part of a fix is to move the reference update notification
after writing the change status to the database. This will have
the least surprising semantics for plugins and any other code
wired into the listener. Commits visible in the reference will
now show as merged if the listener queries for them.

Change-Id: Ib8d034c3f2eb79716b5ebc1ed426a0e8dbb0861a

Former-commit-id: 577738dbdf137c82b896fde32746a51be7beb274