Disable "save" button while saving commit message edit
After editing the commit message and pressing "save", the button
remains active until the message is saved and the dialog is closed.

Usually the edit completes quickly, but on some systems where the
network is slow it can take a few seconds.  If the user presses the
"save" button again (perhaps more than once) it will result in more
than one edit request being sent.  This can in some cases result in
inconsistency between the database and the git repository, which
prevents the change from being viewed any more until a new patch set
is manually uploaded.

Disable the "save" button after it has been pressed, to prevent this.

Change-Id: I820095a1d91b7fd7c498c7c24fa5e7e81e40a7eb

Former-commit-id: 218b15778e7e39fc1bd7dcc9815868e8f3770ab3