Allow http and ssh plugins to provide DynamicItem
The plugin guice environment already allows an HttpPlugin or SshPlugin
to provide a DynamicSet or DynamicMap, but not a DynamicItem.  Add the
missing DynamicItem support.

We only plan to use this soon for http plugins (to allow replacing the
WebSession implementation).  Adding it for ssh plugins while at it
makes the code more symmetrical.

Change-Id: I27c8eb283b1a066f069b28d214788706183dffba

Former-commit-id: c9f7a489bc34a4f6fd1070198c61ec8ae8fc40a8