Populate 'My' menu from user preferences
Instead of hard coding the 'My' menu on client side the menu will now
be populated from the user preferences which are returned from the
server. On server side the menu items are still hard coded, but
follow-up changes will make it configurable per user. The idea is that
users should be able to configure custom dashboards that they want to
link under the 'My' menu.

Change-Id: I6aec34045777114027e5f3f3d37102f67b1036ed
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: 7941a187c3fe3485762c140bed90b5ac1b57209b