Enable notedb tests for LabelTypeIT
Fix construction of labels in ChangeJson to use
ApprovalsUtil.byPatchSet where appropriate, to get appropriate label
copying and normalization behavior.

For now, keep the existing behavior of showing reviewers in the
detailed label output if they have ever voted on any patch set of the
change, even if they are not currently listed as a reviewer.

Change-Id: I1ea2b4110fb3eca41c9bb406e8ee11e1f60a8e69

Former-commit-id: a06b46982b99a1e494f5ded0bfcbf2d88548c716