Allow users to configure the 'My' menu from the preferences screen
Users can now configure in the preferences UI which menu items they
want to see in the 'My' menu. With this change users can't control the
order of the menu items yet. This will be done in a follow up change.

The listAdd icon is taken from the Tango Icon Library [1].

[1] http://tango.freedesktop.org/Tango_Icon_Library

Change-Id: If945cf33fc95430fbd381ae9dd22ce231573e4da
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: 21afc92de45a0dd296f102e0066b49c7cd0b6664