Add a "projects" field for searching projects by prefix
Index alongside the "project" field, which is an exact match field,
since we do not assume index implementations can search exact match
and prefix on the same field. We do not want to modify the existing
"project" to return prefixes.

Upgrade Lucene to 4.7.0, as this was released since the last schema
change.

Change-Id: I7c3379c087fc54af3e5790cb875a5e676e674338

Former-commit-id: 4aed07b7ce7aff1c6588f6e7793e13518b18ad3c