Allow admins to configure the default 'My' menus for all users
If a user hasn't configured any 'My' menu items it falls back to the
menu items that the administrator has configured in the
'refs/users/default' branch in the 'All-Users' projects. If no
menu items are configured there it falls back to the standard menu
items which are hard coded in the server.

Change-Id: I922f68dd6e2effaca96ea2b351c35686a0cc4410
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: 95c3addd1aa48a1d44d8d054b4425cecacdeb189