Refactor MergeabilityChecker to use a builder pattern
This class was growing with various permutations of change, project,
branch, sync/async, which are suited to constructing with a Builder.
We also want to add a priority flag to certain checks to support more
fine-grained scheduling.

In general, due to the strategy of kicking off async work that can
fan out and ignoring errors, try to log errors as they happen instead
of depending on callers to call get() on the resulting Futures.

Change-Id: Ia575793e5ceb5aa3148b45a9a49e4e10e50a5108

Former-commit-id: b49726e145fb55bfdd87ec98c7f2135b9ebbbe57