Split mergeability checks by priority
When a project config is updated or a change is merged, all open
changes on an entire project or branch need to be rechecked for
mergeability. For projects or branches with lots of open changes,
this can flood the queue with potentially slow checks. This will
block simpler cases, such as checking mergeability of a single change
during ReceiveCommits.

Define a new Priority enum on the MergeabilityChecksExecutor so we
can split the queue based on priority, allowing smaller, interactive
changes such as pushes to not wait for the entire queued backlog.

Change-Id: I57f9eb33dbbcd8a560adfe4458ce48d7b0817957

Former-commit-id: 6ea964a0a17b95051db8a56c05a99fcabd14b0f0