Limit archives to only enabled formats
Refactor the configuration of enabled archives to be known to the
GetArchive REST API handler so it can allow only formats that have
been enabled by the administrator.

Use the order listed in download.archive to configure the web
interface, with no formats enabled if the list is empty.

Default to all archive formats enabled, matching the schemes.

Change-Id: I29436f13bbc6fd4a20a7d9151e7f47186f5fbf4b
Former-commit-id: 6edde31009d065b2fcd29834fc90290cd3743e8c