Fix link behaviour inconsistencies in change info panel
In the change info panel there are links on the project name, branch
name, and topic name.  They all behave differently:

 project: opens the project admin page
 branch: opens a query for changes on the branch and topic
 topic: opens a dialog to edit the topic name

Also, next to the project is a query icon that opens the project
default dashboard which is a query for recent changes.

These inconsistent links reduce the useability of the UI.  With this
change, they are made more consistent:

 project: opens the project default dashboard
 branch: opens a query for changes on the branch (without topic)
 topic: opens a query for changes on the topic

The project query icon is replaced with a settings icon that opens
the project admin page.  The topic can still be edited by clicking
on the edit icon.

Change-Id: I09c36bf80949efd8a6237669c8284275c3372d9c
Former-commit-id: 7b83f59157de293bff7904d74d412b6d95f7a7d5