Make gitweb prompt for authorization
Make gitweb return a 401 UNAUTHORIZED response if the user isn't
currently logged in and the project can't be found.

This response is uniform to cover both the cases where anonymous
doesn't have access to the project, or if the project just doesn't
exist.

If the user is authorized then a 404 is continued to be returned.

Bug: Issue 2595
Change-Id: I199a725fc3ec73e3493cadb6ccf2d7ad54262a2e

Former-commit-id: fd7df89ff9045fd5d5f87506c7d214653729cf4e