Fix caching of configs in ProjectState
ProjectState is caching project config files, but it only used the
file name as key and not the ref, which is an issue if the same config
file appears in different refs. This is e.g. the case for the
preference.config file in which the user preferences are stored as it
exists in each user ref. As result most users were getting the wrong
preferences for the 'My' menu entries.

Change-Id: Ib557ff3520bae1a8ee40e3ba6f1b3a8fdd3f880a
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: ecdc64f98bd2118152651d2b596537b1a0a24368