Set uploader to current user in "patchset-created" event upon cherry-picking
When using the web ui (both old and new ChangeScreen) to cherry-pick a
change to a branch that already has this change (e.g.: cherry-picking
on the same branch to get rid of dependencies), the corresponding
patchset-created event had its patchSet.uploader set to the Change's
owner instead of the current user. We now set it to the current user,
so stream-events consumers can properly detect who uploaded the
rebased patch set.

Change-Id: I0613ecae06b7843b8ca0294f83378c3917c26810

Former-commit-id: 980c5bf9ef4f27cc8d25cb0c4aebf8d54b9cc47f