StringListPanel: Do not ignore empty values when adding new row
The position of a value in the value list defines its meaning. If
empty values are ignored the positions of follow-up values are wrong.
Also the up and down arrows for sorting are then displayed in the
wrong columns since they are displayed after the last value of the
value list.

Change-Id: I83a346c48018737ec0154c5f4e224b6de8226801
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: b40f4a157194d7a8503fa127e85d58eac53cb447