Fix potential NPE in queries with --dependencies (ancestors)
If an entry in patch_set_ancestors references a non-existing
entry in the change table, an NPE would be thrown and the query
would be aborted.  Instead, skip over that entry so the query
still works.  This change deals with the case when this happen
looking for ancestors, not neededBys.

Change-Id: I51b23e899895976021f47affa9525548e059845b

Former-commit-id: 683390f7f9cb38c6916e12ba5693b1b3f8bde65a