Fix confusing new change notifications from self added reviewers
If A.U.Thor creates a code review and Re View Er adds themselves
to the change, Gerrit was emailing out the following message:

  Re View Er has uploaded a new change for review.

    https://....

This is confusing and incorrect. Re View Er simply added themselves
as a reviewer to the code review, they did not upload the change.

The message was caused by sending out an email with no reviewers,
which triggered NewChange.vm to use $fromName as the "author" of
the new change. It never should have sent the email as no code
review was requested.

Change-Id: I76808153565ac27a319968525b31d52f72d929b3

Former-commit-id: 788c121b36d207c2d175977abd103144b2c104cb