ChangeScreen2: Only include message on quick approve if reply is open
If the reply box is open include its message if the user clicks
quick approve button instead of post. If the reply box is closed
only send the quick approve button and saved draft comments.

This avoids sending a partial message the user may have started
to draft in the reply box, but thought they aborted by clicking
the cancel button.

The draft message box is now cleared when the user presses the
"Cancel" button in the UI.

Change-Id: Iabdc44a6ce2cf593542183284f4a4760895f346e

Former-commit-id: bf16ed3bb05e3c770d06c693a1ad21c64ec107fd