Add script to find the largest objects in the pack files of a repo
This script is useful if an existing history should be pushed into a
Gerrit project but the push fails because the server has a max object
size limit configured and the existing history contains objects that
are too large. The script helps to identify the large objects so that
they can be removed from the history by using the 'git filter-branch'
command, or so that the limit can be increased.

Change-Id: I99d21d76ac204bf9b53fd589c0ffcbeee2e02422

Former-commit-id: 2d1c219aa3c28b8288f2f2b9506387939f631a01