InlineEdit: Add container to represent change edit
Change edits are created on top of patch sets and are virtual patch
sets that don't exist in the database. A max of one edit per user per
change is stored on ref:

  refs/users/UU/UUUU/edit-CCCC

for edit by user UUUU on top of patch set P of CCCC, where user account
is sharded in the same way as for refs/changes.  The parent of the edit
is SHA-1 of the patch set P it is based on.

Change-Id: I736f5b954ec539a453cf0500a7bd97547bb494d7

Former-commit-id: 8d3c257e68ff8939a19024feaa5476ba88c29d53