ChangeScreen2: Make project settings gear icon a proper hyperlink
Wrapping the icon in an <a href> tag allows users to ctrl-click the
icon to open the settings page in a new tab, instead of replacing the
current change.

If the user clicks in the tab, the icon still consumes the event so
the standard delayed updating token process is used for navigation.

Change-Id: I0c68d3ac7a65ac9885056c83c51d2bfc9ada0533

Former-commit-id: 8398359816e9a1fcea1b5c02b11110e7f48a4141