Add documentation with screenshots for the new review UIs
The new review UI documentation should describe the new change screen
and the new side-by-side diff screen. It provides screenshots in which
the described UI elements are highlighted so that users can easily
find the described functionality in the UI. This documentation is
based on the description of the new screens in the 'What’s new in
Gerrit 2.9' presentation that was shown at the Gerrit User Summit
2014. With this change only the commit message block and the commit
info block are documented. Follow-up changes will document the other
UI elememts. Once the new documentation is ready it should replace
new change screen introduction that is already contained in the
documentation. This guide is purely text-base and hence more difficult
to consume for users.

[1] http://storage.googleapis.com/gerrit-talks/gerrit-new-features-2.9.html

Change-Id: I5dcae829b091d405b9cd6774d63dba869b059f75
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: 4fd84c9051f80059f159c1e5cf47311a67749e11