SideBySide2: Fix B at 100% when A is hidden
Chrome 34.0.1847.131 on Mac didn't give B 100% after hiding A unless
we also let the column header take 100% width.  Previously I had
tested this on Chrome on Linux, but that may have been a different
version of Chrome.

Set both rows to use 100% width on B when A is hidden, making the
layout more explicit to the browser.

Change-Id: I5e334f072563754eecacfe2e514c7987a2dd97b3

Former-commit-id: d1428c154b3b9be23e42b18007d7ddb7cdc4a734