Review UI Docs: Remove note about double-click not adding new comments
It is already obvious from the documentation that double-click is used
to select a word for commenting. The possibilities to add new comments
are clearly described. This is why there is no need to explicitly
mention that double-click is not anymore inserting new comments.

This keeps the documentation clean without any references to how it
was in the old screens.

Change-Id: Ie9f16bde66aee1f1533e30945883190e1265fa24
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: d063df9bdbdc540bbadffe1dfabb7b13d384edad