Keep new draft comment expanded if expand all comments is set
If a new draft comment is added it is automatically collapsed if its
message is lesser than 70 characters, but if the 'Expand All Comments'
diff preference is set to 'Expand' it should stay expanded. The
current behaviour is inconsistent because adding a new draft comment
collapses the comment and after refreshing the page it gets expanded.

Change-Id: Iac6172ea99b17142f357bfdbc04a5f6ca56c602d
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: f3ef2675655f78cb7bb19be97a5f78807a27162b