Velocity: Enable strict reference mode
Velocity 1.6 introduces a concept of strict reference mode which makes
it easier to find misspelled variables or other mistakes in velocity
templates.  Turn it on.

In strict reference mode any attempt to expand a variable that was
not set or put() throws an exception.  In practice this shouldn't
require any changes except that the syntax to check whether a variable
is set is

	#if ($foo)

and not, say,

	#if ($foo != 0)

Change-Id: Id344fa76c769345f84e605984412c51f409d392a

Former-commit-id: 9fd935b7089ab1fda177df2c5f7f46b4c0b3cc5b