Create RebuildNoteDb to migrate data from ReviewDb to notedb
RebuildNotedb is responsible for the migration of pre-existing changes in
the ReviewDb to the notedb. It asynchronously rebuilds every change in
the ReviewDb and all of the related pieces of metadata that can be
stored in the notedb (i.e. PatchSets and PatchSetApprovals). As we add
the ability to store more data in the notedb, this can be updated to
migrate that data as well.

Change-Id: If0fb123630a08ef514611944e7bd0ab1d9e48f80

Former-commit-id: 3f4a1748b22003539e797d7ed36dc41535e1b0b2