Highlight merge commit on change screen
For a reviewer it is important to notice that he is reviewing a merge
commit. Normally this should be obvious by the commit message, but
there may be merge commits with bad commit messages. Highlight merge
commits by displaying an icon for them.

The icon is taken from the Tango Icon Library [1].

[1] http://tango.freedesktop.org/Tango_Icon_Library

Change-Id: I510909a3b146f448492ee11da3c8b80160dec715

Former-commit-id: 69263d352009138d1e27d20909d855eb23fdffac