Documentation: Fix REST API example box width
In 984ed978fb7b40c4c695a8737799aa05b8bf594d [formerly 3eb809dd034cc93685f36a175f52ff7e49e03f23] the max width of the page
was set to 900px, but the REST API example boxes were still set with
a fixed width of 100em which caused them to extend out to the right
of the page.

Remove the fixed width of the box, and set "overflow: auto".  This
makes the boxes fit within the page width, with a horizontal scroll
bar on the ones that have wider content.

Change-Id: I6f64c314e4eb484d9c5120fd0fd5589fcaa8b77e

Former-commit-id: a83c22236210965f0b2805f8b5501fc145884076