[BugFix] set Last-Modified on cached Documentation resources
When a plugin documentation resource was served from
the resourceCache, the Last-Modified HTTP Header was
not set, causing continue HTTP 200 on the same pages
even if the Browser had a cached copy.

This was due to a missing setLastModified() call 
when caching documentation resources. This has now
been fixed with this change and documentation resources
should be more responsive and scalable.

Change-Id: I3dde4178603f930601544219ea6835858384d80b
Former-commit-id: 61cd957b52bb9ce4b996e8fdf7888f9d47f19812