Return 304 for not modified SmallResources
When in-memory cached resources are requested using
If-Modified-Since HTTP header, we can then safely
return a 304 without having to transfer the entire
payload over the network.

This change completes the fix [1]: now that we have
the lastModified date into the SmallResource object
we can use it to actually save bandwidth and allow
better client response time.

[1] https://gerrit-review.googlesource.com/#/c/57213/

Change-Id: If268acacc167a2062260008a3b9767e055970d43
Former-commit-id: c06b6c46156028a8a89d4f58674cd432e3c4b239