New review UI: Fix display of large avatar image in user popup panel
Hovering with the mouse over an avatar image opens a user popup panel
that shows additional information for this user. Beside the name and
the email address it is supposed to show a larger version of the
avatar image. This is important since the avatar images in the new
review UI (new change screen + new side-by-side diff screen) are very
tiny and the persons are difficult to recognize (note that currently
there are no user popups for the avatar images that are displayed for
reviewers and approvals, but only for change messages and inline
comments).

The avatar image in its default size is included in AccountInfo, but
if a different size was requested it didn't load the image from the
server. The problem was likely introduced by commit
da87e49c23a3cab2a0dad75910862e95ac89faea [formerly edb9d2275cfbc5e4905e21786709bce51e7b9e00].

On the old review UI this works properly, so it's a regression with
the new review UI.

Change-Id: Iafa23377aa115ad4dfbe6c0fcd939bcf75b857f7
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: f712cdb27c74d8e19f19e3c15c9bbc32069187d2