Add meta-data to archetype poms
In order to be able to publish the Maven archetypes to Maven Central
the pom files are required to contain elements with meta-data [1].
Having these elements doesn't have any effect on the normal build.

As developers only the maintainers are listed. This is the same what
other projects do, e.g. JGit.

In addition to the required elements also information about the
mailing list and the issue tracker are included.

[1] https://docs.sonatype.org/display/Repository/Sonatype+OSS+Maven+Repository+Usage+Guide#SonatypeOSSMavenRepositoryUsageGuide-6.CentralSyncRequirement

Change-Id: Ia1fd174fcc15d08af44cd8a788ef9cdcb1df5a60
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: cf60c8b65f89666ec0e181906ff5cb7d4b578284